% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_trials.R
\name{bootstrap_trials}
\alias{bootstrap_trials}
\title{Bootstrap Trials}
\usage{
bootstrap_trials(
  numoftrial = 500,
  n,
  v0,
  k_all,
  t_all,
  x_all,
  numofIV,
  r,
  d,
  bw1 = 0.15,
  bw2 = 0.16,
  bw3 = 0.15,
  select_bandwidth = FALSE
)
}
\arguments{
\item{numoftrial}{Integer value. Number of Bootstrap iterations. Defaults to 500.}

\item{n}{Integer value. Sample size, i.e., the number of days.}

\item{v0}{The spot volatilities.}

\item{k_all}{Log-moneyness.}

\item{t_all}{Time-to-maturities.}

\item{x_all}{IV.}

\item{numofIV}{Vector for recording the number of IV observations on each point in time t.}

\item{r}{Risk-free interest rates.}

\item{d}{Dividend yields.}

\item{bw1}{Numeric value for bandwidth of Gamma function estimate in case "select_bandwith" = FALSE. Defaults to 0.15.}

\item{bw2}{Numeric value for bandwidth of Eta^2 function estimate in case "select_bandwith" = FALSE. Defaults to 0.16.}

\item{bw3}{Numeric value for bandwidth of Mu function estimate in case "select_bandwith" = FALSE. Defaults to 0.15.}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}
}
\value{
v: The spot volatilities

Sigma1: The nonparametric bootstrap estimator of the function gamma

Sigma22: The nonparametric bootstrap estimator of the function eta^2

Sigma2: The nonparametric bootstrap estimator of eta

mu_data: The original data of the function mu.

Mu: The nonparametric bootstrap estimator of the function mu.

Rho: The nonparametric bootstrap estimator of the leverage effect function rho.
}
\description{
Generates bootstrap trials for estimating the confidence intervals of an Implied Stochastic Volatility Model
}
