% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_invertible.R
\name{check_invertible}
\alias{check_invertible}
\title{Check Invertibility}
\usage{
check_invertible(m)
}
\arguments{
\item{m}{A NxN matrix.}
}
\value{
Boolean. Outputs TRUE if the matrix is invertible.
}
\description{
Verifies whether the input matrix is invertible.
}
