% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_volatility_state.R
\name{check_volatility_state}
\alias{check_volatility_state}
\title{Check Volatility State}
\usage{
check_volatility_state(mean_impl_vola, target_value, distr = "norm")
}
\arguments{
\item{mean_impl_vola}{the 'mean_impl_vola' output returned by ICC clustering.}

\item{target_value}{an IV value.}

\item{distr}{the distribution based on which the likelihood should be computed. Can be either "norm" for normal or "lognorm" for lognormal.}
}
\value{
0 or 1. Reports to which state the observation belongs.
}
\description{
For a time series clustered by ICC when K=2, this returns to which state an unclustered observation would belong based on the likelihood.
}
