% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ISVM_data.R
\name{data_crypto}
\alias{data_crypto}
\title{Prepare ISVM data}
\usage{
data_crypto(DATA)
}
\arguments{
\item{DATA}{A file generated from the "make_ISVM_data" function}
}
\value{
n: Sample size, i.e., the number of days

beta01: Vector of regression coefficient beta^{(0,1)} (see equation (36))

beta02: Vector of regression coefficient beta^{(0,2)} (see equation (36))

beta10: Vector of regression coefficient beta^{(1,0)} (see equation (36))

v: Vector of spot volatilities, i.e., beta^{(0,0)} in equation (36)

x_all: IV after data filteration

t_all: Time-to-maturities after data filteration

k_all: Log-moneyness after data filteration

r: Risk-free interest rates after data filteration

d: Dividend yields after data filteration

numofIV: Vector for recording the number of IV observations on each day
}
\description{
Preprocessing ISVM data and getting coefficient estimates.
}
