% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ICC_ISVM.R
\name{fit_ICC_ISVM}
\alias{fit_ICC_ISVM}
\title{Fit ICC & ISVM}
\usage{
fit_ICC_ISVM(
  DT_full,
  last_date,
  subset_maturities = TRUE,
  last_only = T,
  N_subs_date = 5,
  K = 2,
  gamma = 10,
  n_boottrial = 500,
  parallel_ICC = FALSE
)
}
\arguments{
\item{DT_full}{A data.table with structure as provided in the example.}

\item{last_date}{When does the sampling period end?}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{last_only}{Should only the last days of the dataset be returned? If set to TRUE this will give exactly as many days as specified in "N_subs_date". Defaults to FALSE.}

\item{N_subs_date}{Number of dates to subset the input data from. Defaults to 5.}

\item{K}{Integer value. Number of clusters. Defaults to 2.}

\item{gamma}{Numeric value for the switching penalty. Defaults to 10.}

\item{n_boottrial}{How many bootstrap iterations should be run?}

\item{parallel_ICC}{Should the ICC algorithm be run in parallel?}
}
\value{
Returns a list with both the ICC and the ISVM result.
}
\description{
Fit an ISV model for each state detected by ICC
}
