% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sub_state_ISVM.R
\name{fit_sub_state_ISVM}
\alias{fit_sub_state_ISVM}
\title{Fit Sub State ISVM}
\usage{
fit_sub_state_ISVM(
  DT_subs_list,
  DT_subs,
  states,
  n_boottrial = 500,
  select_bandwidth = F,
  verbose = F,
  subs = 5,
  threshold = 0,
  impute = F,
  state = NA,
  return_np_estimation_pars = FALSE,
  return_lm = FALSE,
  return_lm_boot = FALSE
)
}
\arguments{
\item{DT_subs_list}{A data set generated by the \code{make_subset} function.}

\item{DT_subs}{A data.table with structure as provided in the example.}

\item{states}{A data.table with the states of each observation.}

\item{n_boottrial}{How many bootstrap iterations should be run?}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}

\item{verbose}{Boolean value. Should the function be verbose?}

\item{subs}{Number of dates to subset the input data from. Defaults to 5.}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{state}{The state for which the model should be calculated}

\item{return_np_estimation_pars}{Should the fitted estimation parameters be returned? Defaults to \code{FALSE}.}

\item{return_lm}{Should the fitted object be returned? Defaults to \code{FALSE}.}

\item{return_lm_boot}{Should the bootstrap fitted object be returned? Only works if \code{return_lm = FALSE}. Defaults to \code{FALSE}.}
}
\value{
Output: An ISVM model for the sub state
}
\description{
Helper function to fit ISVM for single states
}
