% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters.R
\name{get_clusters}
\alias{get_clusters}
\title{Get ICC cluster.}
\usage{
get_clusters(
  DATA,
  plot_data,
  N_subs_date,
  K = 2,
  gamma = 0,
  w_max = 3,
  money_range,
  impute,
  subset_maturities,
  DEBUG = 0,
  random_seed = 123,
  make_beep = F,
  save_plot = T,
  data_only = T,
  distance_function = NA,
  print_plot = TRUE
)
}
\arguments{
\item{DATA}{Output 'train_data' generated from "make_subset" function.}

\item{plot_data}{Output 'plot_data' generated from "make_subset" function.}

\item{N_subs_date}{Number of dates to subset the input data from.}

\item{K}{Integer value. Number of clusters. Defaults to 2.}

\item{gamma}{Numeric value for the switching penalty. Defaults to 0.}

\item{w_max}{Integer value. The number of maximum iterations for ICC. Defaults to 3.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf).}

\item{impute}{Boolean value. Should missing observations be imputed?}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered?}

\item{DEBUG}{Integer value. Parameter for verbosity of ICC algorithm that can be either 0 or 1. Defaults to 0.}

\item{random_seed}{Numeric value for reproducibility of ICC results. Defaults to 123.}

\item{make_beep}{Makes a beep when ICC calculation is finished. Defaults to FALSE.}

\item{save_plot}{Boolean. Should the plots be saved?}

\item{data_only}{Boolean. If set to TRUE, only the data is returned and plotting is skipped. Defaults to FALSE.}

\item{distance_function}{Character value. Which distance should be used? Choose either "euclidean","loglik" (loglikelihood), or "mahalanobis".}

\item{print_plot}{Boolean. Should the plots be saved?}
}
\value{
DATA: The input data.

date: The date of the last observation.

symbols: List of all symbols that were used for clustering.

price_plot: Plot of the clustered price of the underlying index.

vola_plot: Plot of the clustered input data.

states: Time series with the states at each point in time.

mean_impl_vola: Table with mean and sd of the detected clusters.

w_max: Number of maximum iterations specified.

w: Number of actual iterations.

K: The input number of clusters .

scale_gamma: Indicates if the penalty gamma got scaled down.

gamma_init: Initial value provided for gamma.

gamma_value: Actual value for gamma. In case gamma hasn't been scaled down this will be identical to 'gamma_init'.
}
\description{
Computes clusters generated by ICC and returns a list with ICC output data & plots and saves the files if needed.
}
