% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_np_estimation.R
\name{get_np_estimation}
\alias{get_np_estimation}
\title{Get nonparametric boostrap ISVM estimate.}
\usage{
get_np_estimation(
  DT_ISVM,
  boottrials = 500,
  select_bandwidth = F,
  return_np_estimation_pars = FALSE,
  return_lm = FALSE,
  return_lm_boot = FALSE,
  save = FALSE,
  filename = NULL
)
}
\arguments{
\item{DT_ISVM}{A file generated from the "make_ISVM_data" function}

\item{boottrials}{Number of Bootstrap iterations. Defaults to 500.}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}

\item{return_np_estimation_pars}{Should the fitted estimation parameters be returned? Defaults to \code{FALSE}.}

\item{return_lm}{Should the fitted lm object be returned? Defaults to \code{FALSE}.}

\item{return_lm_boot}{Should the bootstrap fitted lm object be returned? Only works if \code{return_lm = FALSE}. Defaults to \code{FALSE}.}

\item{save}{Should the plots be saved? Defaults to FALSE.}

\item{filename}{If "save" = TRUE, what is the filename of the file that should be saved?}
}
\value{
Returns a list with the plots of the nonparametric function estimates.
}
\description{
Getting a Stochastic Volatility estimate with bootstrapping confidence intervals.
}
