% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_states.R
\name{get_sub_states}
\alias{get_sub_states}
\title{Make subset state list}
\usage{
get_sub_states(DATA)
}
\arguments{
\item{DATA}{The output "states" from the "get_clusters" function.}
}
\value{
A list with all states.
}
\description{
Generates a sorted list with all states in provided ICC data.
}
