% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SV_estimate.R
\name{get_SV_estimate}
\alias{get_SV_estimate}
\title{Get SV estimate}
\usage{
get_SV_estimate(DATA)
}
\arguments{
\item{DATA}{A file generated from the "make_ISVM_data" function}
}
\value{
v: The spot volatilities

sigma1: The data of the function gamma (see equation (20))

Sigma1: The nonparametric estimator of the function gamma

sigma22: The data of the function eta^2 (see equation (25))

Sigma22: The nonparametric estimator of the function eta^2

Sigma2: The nonparametric estimator of eta

mu: The data of the function mu (see equation (26))

Mu: The nonparametric estimator of the function mu

Rho: The nonparametric estimator of the leverage effect function rho (see equation (35))

dsigma1: The local linear estimator of the first order derivative of gamma (see equation (24))
}
\description{
Wrapper for preprocessing ISVM data and getting a Stochastic Volatility estimate.
}
