% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_test_result.R
\name{get_test_result}
\alias{get_test_result}
\title{Get test result}
\usage{
get_test_result(
  DATA,
  K = 2,
  gamma = 0,
  w_max = 3,
  money_range = c(0.8, 1.2),
  impute = TRUE,
  subset_maturities = FALSE,
  save_plot = F,
  data_only = F,
  distance_function = "loglik",
  print_plot = T
)
}
\arguments{
\item{DATA}{Output list generated from "make_subset" function.}

\item{K}{Integer value. Number of clusters. Defaults to 2.}

\item{gamma}{Numeric value for the switching penalty. Defaults to 0.}

\item{w_max}{Integer value. The number of maximum iterations for ICC. Defaults to 3.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to FALSE.}

\item{save_plot}{Boolean. Should the plots be saved? Defaults to FALSE.}

\item{data_only}{Boolean. If set to TRUE, only the data is returned and plotting is skipped. Defaults to FALSE.}

\item{distance_function}{Character value. Which distance should be used? Choose either "euclidean","loglik" (loglikelihood), or "mahalanobis". Defaults to "loglik".}

\item{print_plot}{Boolean. Should the plots be saved? Defaults to TRUE.}
}
\value{
DATA: The input data.

date: The date of the last observation.

symbols: List of all symbols that were used for clustering.

price_plot: Plot of the clustered price of the underlying index.

vola_plot: Plot of the clustered input data.

states: Time series with the states at each point in time.

mean_impl_vola: Table with mean and sd of the detected clusters.

w_max: Number of maximum iterations specified.

w: Number of actual iterations.

K: The input number of clusters .

scale_gamma: Indicates if the penalty gamma got scaled down.

gamma_init: Initial value provided for gamma.

gamma_value: Actual value for gamma. In case gamma hasn't been scaled down this will be identical to 'gamma_init'.
}
\description{
ICC clustering for a subset of datapoints. Wrapper for "get_clusters" function.
}
