% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC_R.R
\name{ICC_R}
\alias{ICC_R}
\title{ICC in R}
\usage{
ICC_R(
  DATA,
  K = 2,
  gamma = 0,
  delta = 0,
  w_max = 3,
  DEBUG = 0,
  distance_function = "mahalanobis",
  V = 2.3,
  random_seed = 123
)
}
\arguments{
\item{DATA}{A NxN correlation matrix.}

\item{K}{Integer value. Number of clusters. Defaults to 2.}

\item{gamma}{Numeric value for the switching penalty. Defaults to 0.}

\item{w_max}{Integer value. The number of maximum iterations for ICC. Defaults to 3.}

\item{DEBUG}{Integer value. Parameter for verbosity of ICC algorithm that can be either 0 or 1. Defaults to 0.}

\item{distance_function}{Character value. Which distance should be used? Choose either "euclidean","loglik" (loglikelihood), or "mahalanobis". Defaults to "loglik".}

\item{V}{Parameter V in case "loglik" distance function has been chosen. Defaults to 2.3.}

\item{random_seed}{Numeric value for reproducibility of ICC results. Defaults to 123.}
}
\value{
Vector of ICC state per input observation.

w. Number of successful iterations.
}
\description{
R wrapper for the Python ICC function.
}
