% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_preprocess.R
\name{load_and_preprocess}
\alias{load_and_preprocess}
\title{Load and preprocess}
\usage{
load_and_preprocess(
  DATA,
  filename,
  convert_from_epoch_date = TRUE,
  scale_iv = FALSE
)
}
\arguments{
\item{DATA}{An input data path.}

\item{filename}{The name of the file to be generated.}

\item{convert_from_epoch_date}{Should the data be converted to POSIXct from Epoch/UNIX timestamp? Defaults to TRUE.}

\item{scale_iv}{Should IV be scaled down? Do this when IV values are e.g. 80 instead of 0.8. Defaults to FALSE.}
}
\value{
Outputs preprocessed data.table.
}
\description{
Loads and preprocesses data from the specified input path.
}
