% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_CI.R
\name{make_CI}
\alias{make_CI}
\title{Make confidence interval}
\usage{
make_CI(par_pre, par_sd)
}
\arguments{
\item{par_pre}{A vector with values.}

\item{par_sd}{The standard deviation of the vector.}
}
\value{
List with upper ("u") and lower ("d") confidence interval for provided value.
}
\description{
Generates confidence intervals during ISVM calculation
}
