% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_columns_equal.R
\name{make_columns_equal}
\alias{make_columns_equal}
\title{Make columns equal}
\usage{
make_columns_equal(DATA, mod_cols = c("IV", "money", "mtr", "cp", "r"))
}
\arguments{
\item{DATA}{A data.table.}

\item{mod_cols}{The columns that shpuld be checked for common values.}
}
\value{
Returns the data.table with equal length columns.
}
\description{
Makes sure all provided input columns are of same length based on common entries.
}
