% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_directories.R
\name{make_experiment_directory}
\alias{make_experiment_directory}
\title{Make directories}
\usage{
make_experiment_directory(
  N_subs_date,
  K,
  gamma,
  impute,
  money_range,
  subset_maturities
)
}
\arguments{
\item{N_subs_date}{Number of dates the experiment data was subsetted from.}

\item{K}{Integer value. Number of clusters in the experiment.}

\item{gamma}{Numeric value for the switching penalty in the experiment.}

\item{impute}{Boolean value. Were missing observations imputedin the experiment?}

\item{money_range}{Vector with lower and upper limit of moneyness values that were considered in the experiment.}

\item{subset_maturities}{Boolean value. Was only a subset of the closest maturities considered in the experiment?}
}
\value{
dir_name_vola: The name of the directory for the ICC plots.

dir_name_price:  The name of the directory for the price plots.
}
\description{
Creates directories for saving plots and experiment results with parameters for reproducibility
}
