% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_high_low_state.R
\name{make_high_low_state}
\alias{make_high_low_state}
\title{Make High and Low State}
\usage{
make_high_low_state(DATA, low_state, high_state)
}
\arguments{
\item{DATA}{The input data. Must be a data.table}

\item{low_state}{The state with the lower mean. Either 0 or 1.}

\item{high_state}{The state with the higher mean. Either 0 or 1.}
}
\value{
Boolean. Outputs TRUE if the matrix is invertible.
}
\description{
For a time series clustered by ICC when K=2, this appends to the input data which state is the low and which state is the high mean volatility state.
}
