% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ISVM_data.R
\name{make_ISVM_data}
\alias{make_ISVM_data}
\title{Make subset data}
\usage{
make_ISVM_data(
  ICC_DATA,
  VOLA_DATA,
  start_time = "08:20:00",
  end_time = "07:59:59",
  subset_maturities = FALSE,
  N_subs_maturities = 3,
  money_range = c(0.8, 1.2),
  subs = 5,
  impute = TRUE,
  verbose = FALSE,
  threshold = 0.66,
  state = c(0, 1)
)
}
\arguments{
\item{ICC_DATA}{A list with an object returned from the function "get_clusters".}

\item{VOLA_DATA}{A data.table with structure as provided in the example.}

\item{start_time}{When does the trading day start? Defaults to "08:20:00" (Deribit options expire at "08:00:00").}

\item{end_time}{When does the trading day end? Defaults to "07:59:59" (Deribit options expire at "08:00:00").}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to FALSE}

\item{N_subs_maturities}{Integer value. If 'subset_maturities' = TRUE this will select how many of the closest maturities should be considered. Defaults to 3.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{subs}{Number of dates to subset the input data from. Defaults to 5.}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{verbose}{Boolean value. Should the function be verbose?}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{state}{Vector with unique states. Defaults to c(0,1).}
}
\value{
cp_flag: 0 for call options and 1 for put options.

d: Daily dividend yields in percentage.

date: Numeric dates. One can use the Matlab function datestr to interpret these dates in traditional form.

IVobs: IV observations.

Money: Log-moneyness.

Mtr: Days-to-expiration.

price: Option prices.

r: Risk-free interest rates in percentage.

Sobs: Number of observations of per point in time.
}
\description{
Generates a subset data list that can be used for ISVM calculation.
}
