% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ISVM_grid_list.R
\name{make_ISVM_grid_list}
\alias{make_ISVM_grid_list}
\title{Make ISVM grid list}
\usage{
make_ISVM_grid_list(ISVM_list, variable)
}
\arguments{
\item{ISVM_list}{List of ISVM plots}

\item{variable}{The function for which the plots should be generated}
}
\value{
Returns a list of plots
}
\description{
Helper function for generating ISVM plots
}
