% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ISVM_matrix.R
\name{make_ISVM_matrix}
\alias{make_ISVM_matrix}
\title{Make ISVM matrix}
\usage{
make_ISVM_matrix(
  DATA,
  start_time = "08:20:00",
  end_time = "07:59:59",
  start_date = NA,
  end_date = NA,
  N_subs_date = 5,
  subset_maturities = TRUE,
  N_subs_maturities = 3,
  money_range = c(0.8, 1.2),
  verbose = FALSE,
  impute = TRUE,
  threshold = 0.66,
  value_var
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{start_time}{When does the trading day start? Defaults to "08:20:00" (Deribit options expire at "08:00:00").}

\item{end_time}{When does the trading day end? Defaults to "07:59:59" (Deribit options expire at "08:00:00").}

\item{start_date}{When does the sampling period start? Defaults to NA.}

\item{end_date}{When does the sampling period end? Defaults to NA.}

\item{N_subs_date}{Number of dates to subset the input data from. Defaults to 5.}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{N_subs_maturities}{Integer value. If 'subset_maturities' = TRUE this will select how many of the closest maturities should be considered. Defaults to 3.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{verbose}{Boolean value. Should the function be verbose?}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{value_var}{The variable for which we want to generate a clustering data set.}
}
\value{
Returns a matrix from "make_subset" function without timestamp.
}
\description{
Helper function for wrapping output of "make_subset" function for ISVM as matrix.
}
