% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_data.R
\name{make_plot_data}
\alias{make_plot_data}
\title{Make plot data}
\usage{
make_plot_data(DATA)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}
}
\value{
Returns a data.table with the price of the underlying index.
}
\description{
Helper function for generating plot data for the underlying index.
}
