% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subs_data.R
\name{make_subset}
\alias{make_subset}
\title{Make subset data}
\usage{
make_subset(
  DATA,
  DATE = NA,
  start_time = "08:20:00",
  end_time = "07:59:59",
  N_subs_date = 2,
  threshold = 0.66,
  value_var = "mark_iv",
  money_range = c(0.8, 1.2),
  subset_maturities = TRUE,
  N_subs_maturities = 3,
  min_maturity = 15,
  max_maturity = 60,
  verbose = TRUE,
  impute = FALSE
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{DATE}{A vector of dates to subset from.}

\item{start_time}{When does the trading day start? Defaults to "08:20:00" (Deribit options expire at "08:00:00").}

\item{end_time}{When does the trading day end? Defaults to "07:59:59" (Deribit options expire at "08:00:00").}

\item{N_subs_date}{Number of dates to subset the input data from.}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{value_var}{The variable for which we want to generate a clustering data set.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{N_subs_maturities}{Integer value. If 'subset_maturities' = TRUE this will select how many of the closest maturities should be considered. Defaults to 3.}

\item{min_maturity}{Integer value. Optional and only applies if subset_maturities = FALSE. The lower boundary of maturities to be considered in days. Defaults to 15 days.}

\item{max_maturity}{Integer value. Optional and only applies if subset_maturities = FALSE. The upper boundary of maturities to be considered in days. Defaults to 60 days.}

\item{verbose}{Boolean value. Should the function be verbose?}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}
}
\value{
plot_data: A dataset for plotting the price of the underyling index.

train_data. A dataset for calculating ICC or ISVM values.
}
\description{
Generates a subset data list that can be used for ICC and ISVM calculation.
}
