% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_subs_list.R
\name{make_subset_list}
\alias{make_subset_list}
\title{Make subset data list}
\usage{
make_subset_list(
  DATA,
  N_subs_date = 5,
  start_date = as.Date("1970-01-01"),
  subset_maturities = TRUE,
  N_subs_maturities = 3,
  threshold = 0.66,
  money_range = c(0.8, 1.2),
  verbose = TRUE,
  impute = FALSE,
  parallel = TRUE,
  n_cores = detectCores() * 0.5,
  last_only = F
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{N_subs_date}{Date value. Number of dates to subset the input data from. Defaults to "as.Date("1970-01-01")".}

\item{start_date}{From which day on should a list of subsets be generated?}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{N_subs_maturities}{Integer value. If 'subset_maturities' = TRUE this will select how many of the closest maturities should be considered. Defaults to 3.}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{verbose}{Boolean value. Should the function be verbose?}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{parallel}{Boolean. Should the results be computed in parallel? Won't work on Windows machines. Defaults to TRUE.}

\item{n_cores}{Number of cores. This only applies if parallel = TRUE. Defaults to 50\% of detected cores.}

\item{last_only}{Should only the last days of the dataset be returned? If set to TRUE this will give exactly as many days as specified in "N_subs_date". Defaults to FALSE.
#'}
}
\value{
plot_data: A dataset for plotting the price of the underyling index.

train_data. A dataset for calculating ICC or ISVM values.
}
\description{
List wrapper for "make_subset_data" function that can be used for ICC and ISVM calculation.
}
