% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_estimation_pars.R
\name{np_estimation_pars}
\alias{np_estimation_pars}
\title{Nonparametric ISVM estimation}
\usage{
np_estimation_pars(DT_ISVM, select_bandwidth = F, make_plot = T)
}
\arguments{
\item{DT_ISVM}{A file generated from the "make_ISVM_data" function.}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}

\item{make_plot}{Boolean value. Should the results be plotted? Defaults to TRUE.}
}
\value{
beta01: Vector of regression coefficient beta^{(0,1)} (see equation (36))

beta02: Vector of regression coefficient beta^{(0,2)} (see equation (36))

beta10: Vector of regression coefficient beta^{(1,0)} (see equation (36))

v: Vector of spot volatilities, i.e., beta^{(0,0)} in equation (36)

sigma1: The data of the function gamma (see equation (20))

Sigma1: The nonparametric estimator of the function gamma

sigma22: The data of the function eta^2 (see equation (25))

Sigma22: The nonparametric estimator of the function eta^2

Sigma2: The nonparametric estimator of eta

mu: The data of the function mu (see equation (26))

Mu: The nonparametric estimator of the function mu

Rho: The nonparametric estimator of the leverage effect function rho (see equation (35))

dsigma1: The local linear estimator of the first order derivative of gamma (see equation (24))

n: Sample size, i.e., the number of days

x_all: IV after data filteration

t_all: Time-to-maturities after data filteration

k_all: Log-moneyness after data filteration

r: Risk-free interest rates after data filteration

d: Dividend yields after data filteration

numofIV: Vector for recording the number of IV observations on each day
}
\description{
Getting a nonparametric Stochastic Volatility estimate.
}
