% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(
  DATA,
  N_subs_date = 5,
  subset_maturities = TRUE,
  K = 2,
  N_subs_maturities = 3,
  money_range = c(0.8, 1.2),
  threshold = 0.66,
  parallel = FALSE,
  n_cores = detectCores() * 0.8,
  impute = FALSE,
  verbose = FALSE,
  gamma = 0,
  w_max = 3,
  save_plot = TRUE,
  last_only = FALSE
)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{N_subs_date}{Date value. Number of dates to subset the input data from. Defaults to "as.Date("1970-01-01")".}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{K}{Integer value. Number of clusters. Defaults to 2.}

\item{N_subs_maturities}{Integer value. If \code{subset_maturities = TRUE} this will select how many of the closest maturities should be considered. Defaults to \code{3}.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be c(0,Inf). Defaults to c(0.8,1.2).}

\item{threshold}{How many percent of observations are allowed to be missing for each individual instrument over the observed timeframe? E.g. if set to 0.9, we allow 10\% of observations to be missing. Defaults to 0.66.}

\item{parallel}{Boolean value. Should the clustering be computed in parallel? Won't work on Windows machines. Defaults to \code{FALSE}.}

\item{n_cores}{Number of cores. This only applies if parallel = TRUE. Defaults to 50\% of detected cores.}

\item{impute}{Boolean value. Should missing observations be imputed? Defaults to TRUE.}

\item{verbose}{Boolean value. Should the function be verbose? Defaults to FALSE.}

\item{gamma}{Numeric value for the switching penalty. Defaults to 0.}

\item{w_max}{Integer value. The number of maximum iterations for ICC. Defaults to 3.}

\item{save_plot}{Boolean. Should the plots be saved? Defaults to TRUE.}

\item{last_only}{Should only the last days of the dataset be returned? If set to TRUE this will give exactly as many days as specified in "N_subs_date". Defaults to FALSE.}
}
\value{
DATA: The input data.

date: The date of the last observation.

symbols: List of all symbols that were used for clustering.

price_plot: Plot of the clustered price of the underlying index.

vola_plot: Plot of the clustered input data.

states: Time series with the states at each point in time.

mean_impl_vola: Table with mean and sd of the detected clusters.

w_max: Number of maximum iterations specified.

w: Number of actual iterations.

K: The input number of clusters .

scale_gamma: Indicates if the penalty gamma got scaled down.

gamma_init: Initial value provided for gamma.

gamma_value: Actual value for gamma. In case gamma hasn't been scaled down this will be identical to 'gamma_init'.
}
\description{
Wrapper function to generate ICC estimate for provided input data.
}
