% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICC.R
\name{run_ICC}
\alias{run_ICC}
\title{Run ICC.}
\usage{
run_ICC(
  PATH,
  subs_date_list = c(1),
  gamma_list = c(1),
  money_range = c(0.8, 1.2),
  K = 2,
  parallel = TRUE,
  parallel_preprocessing = TRUE,
  subset_maturities = TRUE,
  N_subs_maturities = 3,
  impute = FALSE
)
}
\arguments{
\item{PATH}{Output 'train_data' generated from \code{make_subset} function.}

\item{subs_date_list}{A vector of dates to generate a subset for clustering. Defaults to \code{ c(1)}.}

\item{gamma_list}{A vector of gamma values for clustering the data. Defaults to \code{ c(1)}.}

\item{money_range}{Vector with lower and upper limit of moneyness values to consider. The natural boundary would be \code{c(0,Inf)}.}

\item{K}{Integer value. Number of clusters. Defaults to \code{2}.}

\item{parallel}{Boolean value. Should the clustering be computed in parallel? Won't work on Windows machines. Defaults to \code{FALSE}.}

\item{parallel_preprocessing}{Boolean value. Should the clustering be run in parallel? Defaults to \code{FALSE.}}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered?}

\item{N_subs_maturities}{Integer value. If \code{subset_maturities = TRUE} this will select how many of the closest maturities should be considered. Defaults to \code{3}.}

\item{impute}{Boolean value. Should missing observations be imputed?}
}
\value{
Returns a list with the results from "get_clusters":

DATA: The input data.

date: The date of the last observation.

symbols: List of all symbols that were used for clustering.

price_plot: Plot of the clustered price of the underlying index.

vola_plot: Plot of the clustered input data.

states: Time series with the states at each point in time.

mean_impl_vola: Table with mean and sd of the detected clusters.

w_max: Number of maximum iterations specified.

w: Number of actual iterations.

K: The input number of clusters .

scale_gamma: Indicates if the penalty gamma got scaled down.

gamma_init: Initial value provided for gamma.

gamma_value: Actual value for gamma. In case gamma hasn't been scaled down this will be identical to 'gamma_init'.
}
\description{
Wrapper function of run_experiment that also loads, preprocesses and saves test data.
}
