% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ISVM.R
\name{run_ISVM}
\alias{run_ISVM}
\title{Run ISVM}
\usage{
run_ISVM(
  ICC_DT_full,
  DT_full,
  subset_maturities = TRUE,
  N_subs_date = 5,
  n_boottrial = 500,
  min_nrow = 25,
  select_bandwidth = F,
  return_np_estimation_pars = FALSE,
  return_lm = FALSE,
  return_lm_boot = FALSE
)
}
\arguments{
\item{ICC_DT_full}{The output object generated by the function \code{run_experiment}.}

\item{DT_full}{A data.table with structure as provided in the example.}

\item{subset_maturities}{Boolean value. Should only a subset of the closest maturities be considered? Defaults to TRUE.}

\item{N_subs_date}{Number of dates to subset the input data from. Defaults to 5.}

\item{n_boottrial}{How many bootstrap iterations should be run? Defaults to 500.}

\item{min_nrow}{Minimum observations required per cluster so that ISVM is calculated. Defaults to 25.}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}

\item{return_np_estimation_pars}{Should the fitted estimation parameters be returned? Defaults to \code{FALSE}.}

\item{return_lm}{Should the fitted object be returned? Defaults to \code{FALSE}.}

\item{return_lm_boot}{Should the bootstrap fitted object be returned? Only works if \code{return_lm = FALSE}. Defaults to \code{FALSE}.}
}
\value{
Returns a list with both the ICC and the ISVM result.
}
\description{
Run the ISVM algorithm
}
