% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_batches.R
\name{separate_batches}
\alias{separate_batches}
\title{Separate batches}
\usage{
separate_batches(DATA, t_diff_lim = 250)
}
\arguments{
\item{DATA}{A data.table with structure as provided in the example.}

\item{t_diff_lim}{Numeric value. Limit after which a new batch is started. Defaults to 250.}
}
\value{
Separates data by batches based on time difference
}
\description{
Helper function for separating batches of observations that are sampled approximately at the same time.
}
