% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVestimate.R
\name{SVestimate}
\alias{SVestimate}
\title{SV estimate}
\usage{
SVestimate(
  v,
  beta01,
  beta02,
  beta10,
  r,
  d,
  date = "",
  bw1 = 0.15,
  bw2 = 0.16,
  bw3 = 0.15,
  select_bandwidth = F,
  nmulti = 10,
  make_plot = TRUE
)
}
\arguments{
\item{v}{Vector of spot volatilities, i.e., beta^{(0,0)} in equation (36)}

\item{beta01}{Vector of regression coefficient beta^{(0,1)} (see equation (36))}

\item{beta02}{Vector of regression coefficient beta^{(0,2)} (see equation (36))}

\item{beta10}{Vector of regression coefficient beta^{(1,0)} (see equation (36))}

\item{r}{Risk-free interest rates.}

\item{d}{Dividend yields.}

\item{date}{Character value. Estimation date for plot title. Optional.}

\item{bw1}{Numeric value for bandwidth of Gamma function estimate in case "select_bandwith" = FALSE. Defaults to 0.15.}

\item{bw2}{Numeric value for bandwidth of Eta^2 function estimate in case "select_bandwith" = FALSE. Defaults to 0.16.}

\item{bw3}{Numeric value for bandwidth of Mu function estimate in case "select_bandwith" = FALSE. Defaults to 0.15.}

\item{select_bandwidth}{Boolean value. Should the optimal bandwidth for the local polynomial kernel regression be selected automatically? If TRUE, bandwidth will be selected according to least-squares cross-validation. Defaults to FALSE.}

\item{nmulti}{How many iterations should be performed for bandwidth selection? Defaults to 10.}

\item{make_plot}{Boolean value. Should the results be plotted? Defaults to TRUE.}
}
\value{
v: The spot volatilities

sigma1: The data of the function gamma (see equation (20))

Sigma1: The nonparametric estimator of the function gamma

sigma22: The data of the function eta^2 (see equation (25))

Sigma22: The nonparametric estimator of the function eta^2

Sigma2: The nonparametric estimator of eta

fit_gamma: The fit of the nonparametric regression of gamma

fit_eta2: The fit of the nonparametric regression of eta^2

fit_mu: The fit of the nonparametric regression of mu

mu: The data of the function mu (see equation (26))

Mu: The nonparametric estimator of the function mu

Rho: The nonparametric estimator of the leverage effect function rho (see equation (35))

dsigma1: The local linear estimator of the first order derivative of gamma (see equation (24))
}
\description{
Getting a Stochastic Volatility estimate.
}
